Remove ssh-keyscan on vagrant up/halt
SSH configuration generated via `vagrant ssh-config` defines a couple
of ssh client options which effectively negate the need to ever do
ssh-keyscan, because ssh client (invoked via docker CLI) will never
check remote ssh key fingerprints.

It might not be the smartest thing in the world, but that's how it is.